Add cfpb- prefix to filenames to avoid confusion
There are some packages that have multiple eponymous files and are
susceptible to naming collisions when bundling. E.g. the buttons
packages has a `src/buttons.less` and a `src/atoms/buttons.less`.

Having the `cf-` prefix fixed this issue but now that we're dropping
the CF brand I replaced it with a `cfpb-` prefix. E.g. the buttons
entrypoint is now `src/cfpb-buttons.less`.