Upgrade to Jekyll 4
This commit updates Jekyll to version 4. Closes #465.

See these pages for documentation on the upgrade:

https://jekyllrb.com/news/2019/08/20/jekyll-4-0-0-released/
https://jekyllrb.com/docs/upgrading/3-to-4/

Minor changes include:

- Tried to use the built-in relative_url filter instead of having
numerous references to site.baseurl.
- Added --profile to 'yarn watch' to make it easier to keep track of
build times.

To test, run 'bundle update', then 'yarn', and 'yarn start'.

You might seem some errors like "Empty `slug` generated for ''" -- these
are because we have some empty variation names and, as of Jekyll 4, the
use of | slugify on these generates a warning. These should be fixed but
that is independent of this upgrade -- note for example the empty
variation names on
https://cfpb.github.io/design-system/foundation/typography.