Fix GH Actions script when no changes exist
Currently we run the gh-pages GitHub Action even if there are no changes
in the documentation website.

There's a line that tries to exit early if there are no changes to
commit, but it doesn't work properly. Because we use 'set -e' at the top
of the script, when 'git commit' returns 1 (if there are no changes to
commit), it causes the script to fail with an error.

For example, that happened on this Actions run:

https://github.com/chosak/design-system/runs/637557200

Note the message right before failure with "nothing to commit, working
tree clean". This also returns a non-zero status code, which causes the
script to fail.

This commit instead first checks to see if there are any changes to
commit, and gracefully exits if not.

An example run where there was a graceful exit:

https://github.com/chosak/design-system/runs/643781788

Note the message "no local changes to commit".

And an example run with a change that was properly committed:

https://github.com/chosak/design-system/runs/643801916