Generalize tabs and "show details" functionality to support CMS previews
In order to have our vanilla JS tab and "show details" toggling code work
with both the built static site and Netlify's React-based preview pane, the
functionality is moved into discrete files and exported as needed.

Gov UK's tab switching is incompatible with Netlify CMS because it modifies the
URL's hash so a dumbed down version is provided soley for the preview pane.