Add link to CMS instructions page for new users
Creates a new admin widget that includes a link to our CMS instructions
page. It's identical to our string widget but includes a link underneath
it.

This approach adds some redundant code but adding/editing widgets is the
sustainable method of modifying the CMS's left-hand editing pane.
The alternative would be forking netlify-cms and tweaking its source code
which is a maintenance headache.