Ensure correct multiselect input labels and IDs
This change properly handles multiselects defined with an `id` but not a
`name`, as in the example in the design system docs [0]:

<select class="o-multiselect" id="something" multiple>

Currently, the JS code generates a text <input> with an empty `id`
attribute. This means that the associated <label> points to the wrong
thing, causing failures in accessibility audits, for example [1].

Additionally, the current behavior fails to ensure that the generated
checkboxes have unique `id`s, simply copying the option `value`s. This
commit prepends each `id` with the enclosing `input` ID, ensuring
uniqueness on the page.

[0] https://cfpb.github.io/design-system/components/dropdowns-and-multiselects#multiselects-1
[1] https://storage.googleapis.com/lighthouse-infrastructure.appspot.com/reports/1607472905780-28501.report.html