Add netlify-cms field for rendered code snippet code
Sometimes we want to add additional JS or HTML to a sample
component's code snippet to help it render properly on the DS website.
The extra code is irrelevant to the user and might confuse them if they
tried to copy/paste the code snippet.

This adds a new optional field that will override the rendered code
snippet but leave raw code snippet untouched.