Only hide expandable content during animation (#1247)
Expandables' overflow is hidden to prevent their contents from spilling
outside their containers when collapsing. A side effect is that UI elements
that *should* be allowed to spill outside their containers  (like multiselect
dropdowns) are unable to do so. The solution is to only hide overflow during
the collapsing animation.