Fix field-level error message margins
Field-level error messages are meant to have a 15px margin between the
field with the error and the message. That margin was only present
when the error message was in an .m-form-field container. Our text
input with button pattern (where the text field and the button are on
the same line) don't have the .m-form-field wrapper, so error messages
in those kinds of forms had no top margin.