Fix broken heading order on heroes page
PR 1299 modified the content of the heroes page and added several
examples with <h1> tags. This breaks Lighthouse accessibility checking,
with then fails with "Heading elements are not in a
sequentially-descending order". To fix this, this commit changes those
<h1> tags to be <p class="h1">.

This was also how the previous hero content was configured in
f3085f2831fc377307987bf0c7b42c56d0063827.