Move breakout card links slightly higher
Increasing the breakout card's font size to h3 causes the text to run
up against the card's bottom border. Instead of adding more spacing
underneath the link and subsequently increasing the height of the entire
card, it's truer to the design to decrease the spacing above the text
and center the text between the image and bottom border.