Consolidate cf.gov responsive table styles, ignore print media
Our responsive table styles are duplicated across cf.gov and the DS.
This commit consolidates them in the DS and a subsequent PR will remove
them from cf.gov.

Also, printing a page triggers our smallest 600px breakpoint, causing
our mobile styles to be applied. This is fine for most of our design patterns
but it provides a degraded experience on pages with HTML tables. The
responsive table media query has been updated to only apply to `screen`
media types and not `print`.