Add hybrid stack-on-small table pattern
Adds a new table pattern that lets users stack some, but not all, columns
on smaller screens. It requires passing a CSS variable in a style attribute
which isn't great. We can hopefully improve the pattern in the future and
remove this requirement.

See https://github.local/Design-Development/Design-Thinking-and-User-Research/issues/232