Add exports config to DS package
The `exports` field allows multiple entrypoints to be defined instead of
just a single `main` entrypoint. This allows the bundled code to be selectively
imported by developers. For our purposes, we can package the majority of our DS
code into a single file but keep the tooltip code separate. Tooltips are used
on very few pages and they rely on a third party library that we don't want users
to unnecessarily download on every page of cf.gov.

See https://nodejs.org/api/packages.html#package-entry-points