Remove unused Lighthouse results upload
A nightly GitHub Action runs Lighthouse against the Design System
documentation site, to ensure that it meets with accessibility
guidelines.

Currently this Action is generating a warning message, for example [0]:

> No files were found with the provided path: .lighthouseci.
> No artifacts will be uploaded.

This warning occurs because the upload step can't find the Lighthouse
results.

We don't actually use the raw Lighthouse results in any way; as part of
the nightly test, HTML reports are uploaded to temporary Google storage
and their public URLs are included in the Action log.

Thus there's no need to upload the raw run results. This commit removes
the upload step, thus fixing the warning.

[0] https://github.com/cfpb/design-system/actions/runs/11524988694