Cleaning up
Based on a review with Ans on Friday there were a few changes suggested.

Fixed up @use in the web component scss so that it uses all of abstracts again. (I initially thought that custom-props scss was still using $ variables. It wasnt)

Refactor button styles for cfpb-button to not be a mixin any longer.

Update index.js in cfpb-button to no longer include cfpb-color-palettes-default-tokens.css because we can just use custom-props from the standard components.

*todo*

Figure out if there is another way to include the btn-default-tokens.css. Right now I have included it in the index.js in order for the styles to be available to the shadow dom. Adding them into cfpb-button.scss with @use or @import is always leaving the css custom properties undefined.