restructuring to have web components use JSON (#2431)
* restructuring to have web components use JSON

duplicated contents of abstracts and base into elements/abstracts and elements/base
updated to using css custom props in base.scss instead of scss vars
added elements/abstracts and elements/base to index.scss
added vars.css to abstracts

this should establish a base to hook up the JSON tokens as the next step

* renaming sizing-vars to not be an scss partial file

As @anselmbradford  pointed out we do not do that for any other files in the codebase, so the usage here was inconsistent.

Renamed to remove the underscore.

* removing unused js files and one grid-gutter-size var

* getting rid of the leftover js files in dist